refactored project code
