--- README.md
@@ -1,14 +1,0 @@
-# Gender Classification based on Author tweets
-
-## This repo contains our code for the gender classification problem based on author's tweets
-
-**Team Members**
-1. Rishabh Katna(17MI552)
-1. Kashish Kalsi(17MI507)
-1. Srajika Gupta(17MI516)
-
-**To run the project, download/clone the repo and run the following commands**
-1. `cd gender_classifier`
-1. `python3 main.py`
-
-> **__NOTE__:** In case of an error, Look at the *config* dictionary and check the path of the xmls directory
